nostr: remove `RawRelayMessage`
