+++ crates/nostr-relay-pool/src/relay/inner.rs
@@ -17,7 +17,6 @@ use async_wsocket::{ConnectionMode, Message};
 use atomic_destructor::AtomicDestroyer;
 use negentropy::{Id, Negentropy, NegentropyStorageVector};
 use negentropy_deprecated::{Bytes as BytesDeprecated, Negentropy as NegentropyDeprecated};
-use nostr::event::raw::RawEvent;
 use nostr::secp256k1::rand::{self, Rng};
 use nostr_database::prelude::*;
 use tokio::sync::mpsc::{self, Receiver, Sender};
@@ -949,24 +948,25 @@ impl InnerRelay {
         }
 
         // Handle msg
-        match RawRelayMessage::from_json(msg)? {
-            RawRelayMessage::Event {
+        match RelayMessage::from_json(msg)? {
+            RelayMessage::Event {
                 subscription_id,
                 event,
-            } => self.handle_raw_event(subscription_id, event).await,
-            m => Ok(Some(RelayMessage::try_from(m)?)),
+            } => {
+                self.handle_event_msg(subscription_id.into_owned(), event.into_owned())
+                    .await
+            }
+            m => Ok(Some(m)),
         }
     }
 
-    async fn handle_raw_event(
+    async fn handle_event_msg(
         &self,
-        subscription_id: String,
-        event: RawEvent,
+        subscription_id: SubscriptionId,
+        event: Event,
     ) -> Result<Option<RelayMessage<'static>>, Error> {
-        let kind: Kind = Kind::from(event.kind);
-
         // Check event size
-        if let Some(max_size) = self.opts.limits.events.get_max_size(&kind) {
+        if let Some(max_size) = self.opts.limits.events.get_max_size(&event.kind) {
             let size: usize = event.as_json().len();
             let max_size: usize = max_size as usize;
             if size > max_size {
@@ -975,7 +975,7 @@ impl InnerRelay {
         }
 
         // Check tags limit
-        if let Some(max_num_tags) = self.opts.limits.events.get_max_num_tags(&kind) {
+        if let Some(max_num_tags) = self.opts.limits.events.get_max_num_tags(&event.kind) {
             let size: usize = event.tags.len();
             let max_num_tags: usize = max_num_tags as usize;
             if size > max_num_tags {
@@ -986,26 +986,27 @@ impl InnerRelay {
             }
         }
 
-        // Deserialize partial event (id, pubkey and sig)
-        let partial_event: PartialEvent = PartialEvent::from_raw(&event)?;
+        // Check if the event is expired
+        if event.is_expired() {
+            return Err(Error::EventExpired);
+        }
+
+        // Check event admission policy
+        if let Some(policy) = self.state.admit_policy.get() {
+            if let AdmitStatus::Rejected = policy
+                .admit_event(&self.url, &subscription_id, &event)
+                .await?
+            {
+                return Ok(None);
+            }
+        }
 
         // Check if event status
-        let status: DatabaseEventStatus = self.state.database().check_id(&partial_event.id).await?;
+        let status: DatabaseEventStatus = self.state.database().check_id(&event.id).await?;
 
         // Event deleted
         if let DatabaseEventStatus::Deleted = status {
             return Ok(None);
-        }
-
-        // Deserialize missing fields
-        let missing: MissingPartialEvent = MissingPartialEvent::from_raw(event)?;
-
-        // Compose full event
-        let event: Event = partial_event.merge(missing);
-
-        // Check if it's expired
-        if event.is_expired() {
-            return Err(Error::EventExpired);
         }
 
         // Check if coordinate has been deleted
@@ -1019,18 +1020,6 @@ impl InnerRelay {
                 return Ok(None);
             }
         }
-
-        // Check event admission policy
-        if let Some(policy) = self.state.admit_policy.get() {
-            if let AdmitStatus::Rejected = policy
-                .admit_event(&self.url, &subscription_id, &event)
-                .await?
-            {
-                return Ok(None);
-            }
-        }
-
-        let subscription_id: SubscriptionId = SubscriptionId::new(subscription_id);
 
         // TODO: check if filter match
