+++ crates/nostr/src/lib.rs
@@ -63,7 +63,7 @@ pub use self::filter::{Alphabet, Filter, SingleLetterTag};
 #[doc(hidden)]
 pub use self::key::{Keys, PublicKey, SecretKey};
 #[doc(hidden)]
-pub use self::message::{ClientMessage, RawRelayMessage, RelayMessage, SubscriptionId};
+pub use self::message::{ClientMessage, RelayMessage, SubscriptionId};
 #[doc(hidden)]
 pub use self::nips::nip01::Metadata;
 #[doc(hidden)]
